Extract Stats business logic to ServiceLayer
